requirements.txt
python-json-logger>=3
pyyaml>=6